changement de variables de Tower pour adaptabilité

+++ src/tower.hpp
@@ -25,5 +25,5 @@ struct Tower{
     void tower_map_apparition();  //méthode pour faire apparaître les sprites des towers sur la map
     void tower_interface_apparition(); //méthode pour faire apparaître en continu les sprites des towers sur la boutique de l'interface
     void tower_aiming();  //La méthode pour détecter un ennemi à partir de fire_distance
-    void tower_fire(int x_pos_enemy, int y_pos_enemy);  //méthode pour faire feu sur un ennemi quand il y a détection
+    void tower_fire(float x_pos_enemy, float y_pos_enemy);  //méthode pour faire feu sur un ennemi quand il y a détection
 };
